==> 1_Data_extraction.ipynb <==
## JSTOR Constellate

**Step 1: Dropping n-grams of the JSTOR file**

Script for removing the n-grams from articles exported from JSTOR Constellate (new version of data).
It opens the JSON-Lines file for each period that contains within it the articles. After processing, saves each article as a file in a general folder.


import pandas as pd
import numpy as np
import json
import os 

os.getcwd()

# Obtain all files in the path
list_files = os.listdir('/Users/carolinachru/Documents/LT/M498_Cap/data/jstor_constellate/all')
list_files

# If UnicodeDecodeError check if we have file .DS_store
# For dropping:
# cd [in past path], then find . -name '.DS_Store' -type f -delete

for file in list_files:
    path = '/Users/carolinachru/Documents/LT/M498_Cap/data/jstor_constellate/' # Path to all files
    main_file = path + 'all/' + file
    
    # Open file
    with open(main_file) as f:
        data0 = [json.loads(line) for line in f]
        
    print('Length of', file, 'is', len(data0))
    
    # Remove n-grams
    data = data0[:]
    for art in data:
        art.pop('bigramCount', None)
        art.pop('trigramCount', None)
        art.pop('unigramCount', None)
    
    # New name
    num = 1
    for art in data:  
        
        # Save file by file in folder
        num_str = str(num)
        name_temp = path + '/all_nograms/' + file + '_' + num_str + '.json'
        
        with open(name_temp, 'w') as fp:
            json.dump(art, fp)
        
        num += 1
