==> 4_Gender_identification.ipynb <==
## JSTOR Constellate

**Step 4: Identifying gender from first names**

Script for identifying gender from first names. Uses the csv file created previously which contains first names for each author.


import pandas as pd
import numpy as np
import json
import os 
import csv

os.getcwd()
path = '/Users/carolinachru/Documents/LT/M498_Cap/data/'
os.chdir(path)

# Package 1: gender
import gender

gd = gender.GenderDetector()
gd.get_gender('Anne John')

p1 = gd.get_gender('Andrea').gender # Saving
p1

# Package 2: genderComputer
from genderComputer.genderComputer import GenderComputer

gc = GenderComputer()
print(gc.resolveGender('Anne John', 'None'))

p2 = gc.resolveGender('Andrea Ferrari', 'Italy') # Saving
p2

os.getcwd()
path = '/Users/carolinachru/Documents/LT/M498_Cap/data/jstor_constellate/final/'
os.chdir(path)

### 1. Open files

# Open firstname's file with base python (use this)
with open('all_authors_firstname.csv', newline = '') as file:
    reader = csv.reader(file)
    data_firstn = list(reader)

data_firstn.pop(0) # Function returns deleted value

# Open all names's file with base python (use this)
with open('all_authors_allname.csv', newline = '') as file:
    reader = csv.reader(file)
    data_alln = list(reader)

data_alln.pop(0) # Function returns deleted value

### 2. Gender identification

print('All names:', data_alln[0])
print('First names:', data_firstn[0])

#data_alln_test = data_alln[0:99] - This for a sample

# P1
gd = gender.GenderDetector()
p1 = gd.get_gender('A').gender
print(p1)

# P2
gc = GenderComputer()
p2 = gc.resolveGender('A', 'None')
p2

data_alln[0]

def extract_gender(alist):
    '''From alist, extracts gender from column number 2.
    Assumes structure of data is ['1', http://www.jstor.org/stable/29788654', 'ANDREA', 'NA', 'NA', ...]'''
    # New data
    data_alln_gender_p1 = [] 
    data_alln_gender_p2 = []

    for art in alist: 
        temp1 = []
        temp2 = []
        
        # Append the article's id
        temp1.append(art[1])
        temp2.append(art[1])

        for index in list(range(2, 64)): # Just names, not id nor the first column
            name = art[index]
            
            if name != 'NA' and str.isspace(name) == False:
                gen_p1 = gd.get_gender(name).gender
                gen_p2 = gc.resolveGender(name, 'None')
            else:
                gen_p1 = 'NA'
                gen_p2 = 'NA'
            
            # Append to cell
            temp1.append(gen_p1) 
            temp2.append(gen_p2)
        
        # Information for packages, append to row (main dataset)
        data_alln_gender_p1.append(temp1) 
        data_alln_gender_p2.append(temp2) 
    
    return data_alln_gender_p1, data_alln_gender_p2


data_gender1_alln, data_gender2_alln = extract_gender(data_alln)
data_gender1_firstn, data_gender2_firstn = extract_gender(data_firstn)

print('Extracted!')

# p1 = gender; p2 = genderComputer
np.savetxt("all_authors_firstn_gender_Pgender.csv", data_gender1_firstn, delimiter=",", fmt='%s')
np.savetxt("all_authors_firstn_gender_PgenderComputer.csv", data_gender2_firstn, delimiter=",", fmt='%s')

np.savetxt("all_authors_alln_gender_Pgender.csv", data_gender1_alln, delimiter=",", fmt='%s')
np.savetxt("all_authors_alln_gender_PgenderComputer.csv", data_gender2_alln, delimiter=",", fmt='%s')

# Open file with pandas - (desuso)
#df0 = pd.read_csv("all_authors_firstname.csv", header = None, skiprows = 1, index_col = 0)
#print(df0[1])
